==> Algo/Project/1_logistics_model.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score, classification_report
from sklearn.preprocessing import StandardScaler
import joblib

# Load the dataset
file_path = 'Phishing_Legitimate_full.csv'
data = pd.read_csv(file_path)

# Separate features and target variable
X = data.drop(columns=['CLASS_LABEL', 'id'])  # Drop 'CLASS_LABEL' and 'id' columns for features
y = data['CLASS_LABEL']  # Target variable

# Check for missing values
if data.isnull().sum().sum() > 0:
    print("The dataset contains missing values.")
else:
    # Standardize the features
    scaler = StandardScaler()
    X_scaled = scaler.fit_transform(X)
    
    # Split the data into training and testing sets
    X_train_scaled, X_test_scaled, y_train, y_test = train_test_split(X_scaled, y, test_size=0.2, random_state=42)
    
    # Create and train the logistic regression model with scaled data
    log_reg_model_scaled = LogisticRegression(max_iter=10000)
    log_reg_model_scaled.fit(X_train_scaled, y_train)
    
    # Make predictions on the test data
    y_pred_scaled = log_reg_model_scaled.predict(X_test_scaled)
    
    # Evaluate the model with scaled data
    accuracy_scaled = accuracy_score(y_test, y_pred_scaled)
    classification_rep_scaled = classification_report(y_test, y_pred_scaled)
    
    # Print the results
    print(f'Accuracy: {accuracy_scaled * 100:.2f}%')
    print('Classification Report:')
    print(classification_rep_scaled)
    
    # Save the model
    model_filename = 'logistic_regression_model.pkl'
    joblib.dump(log_reg_model_scaled, model_filename)
    print(f'Model saved as {model_filename}')
    
    # Save the scaler
    scaler_filename = 'scaler.pkl'
    joblib.dump(scaler, scaler_filename)
    print(f'Scaler saved as {scaler_filename}')
